Move icon methods from ApplicationHelper to IconHelper
